requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
django-filter==2.2.0
djangorestframework==3.10.3
gunicorn==19.9.0
Markdown==3.1.1
Pillow==6.0.0
psycopg2-binary==2.8.3
python-dotenv==0.10.3
pytz==2019.1
sqlparse==0.3.0
